+++ src/canisters/frontend/ic-certified-assets/src/types.rs
@@ -109,6 +109,7 @@ pub struct AssetProperties {
     pub max_age: Option<u64>,
     pub headers: Option<HashMap<String, String>>,
     pub allow_raw_access: Option<bool>,
+    pub is_aliased: Option<bool>,
 }
 
 #[derive(Clone, Debug, CandidType, Deserialize)]
@@ -117,6 +118,7 @@ pub struct SetAssetPropertiesArguments {
     pub max_age: Option<Option<u64>>,
     pub headers: Option<Option<HashMap<String, String>>>,
     pub allow_raw_access: Option<Option<bool>>,
+    pub is_aliased: Option<Option<bool>>,
 }
 
 #[derive(Clone, Debug, Eq, PartialEq, CandidType, Deserialize)]